+++ plugin-route-analysis/src/main/kotlin/com/linecorp/intellij/plugins/armeria/explorer/support/RouteContributor.kt
@@ -1,5 +1,0 @@
-package com.linecorp.intellij.plugins.armeria.explorer.support
-
-internal fun interface RouteContributor {
-    fun collect(context: RouteCollectContext)
-}